delete plug, processing error, caching forecast weather, delete secret api, change http_request, display weather

+++ lib/weither/scheme/weather.ex
@@ -7,7 +7,7 @@ defmodule Weither.Scheme.Weather do
     field :humidity, :integer
     field :pressure, :integer
     field :temp, :float
-    field :wind_speed, :integer
+    field :wind_speed, :float
 
     timestamps()
   end
